refactor: modularize and reorganize managers, models, and serializers
