+++ app/repositories/managers/__init__.py
@@ -1,0 +1,6 @@
+from .base_manager import BaseManager
+from .beverage_manager import BeverageManager
+from .index_manager import IndexManager
+from .ingredient_manager import IngredientManager
+from .order_manager import OrderManager
+from .size_manager import SizeManager
