+++ app/repositories/managers/beverage_manager.py
@@ -1,0 +1,13 @@
+from typing import Sequence
+from .base_manager import BaseManager
+from app.repositories.models.beverage import Beverage
+from app.repositories.serializers.beverage_serializer import BeverageSerializer
+
+class BeverageManager(BaseManager):
+    model = Beverage
+    serializer = BeverageSerializer
+    
+    @classmethod
+    def get_by_id_list(cls, ids: Sequence):
+        return cls.session.query(cls.model).filter(cls.model._id.in_(set(ids))).all() or []
+
