+++ app/repositories/models/__init__.py
@@ -1,0 +1,8 @@
+from app.plugins import db
+from .ingredient import Ingredient
+from .beverage import Beverage
+from .order import Order
+from .order_beverage import OrderBeverage
+from .order_ingredient import OrderIngredient
+from .size import Size
+
