+++ app/repositories/serializers/__init__.py
@@ -1,0 +1,8 @@
+from app.plugins import ma
+from .beverage_serializer import BeverageSerializer
+from .ingredient_serializer import IngredientSerializer
+from .order_serializer import OrderSerializer
+from .order_beverage_serializer import OrderBeverageSerializer
+from .order_ingredient_serializer import OrderIngredientSerializer
+from .size_serializer import SizeSerializer
+
