+++ app/repositories/serializers/ingredient_serializer.py
@@ -1,0 +1,9 @@
+from app.plugins import ma
+from app.repositories.models.ingredient import Ingredient
+
+class IngredientSerializer(ma.SQLAlchemyAutoSchema):
+
+    class Meta:
+        model = Ingredient
+        load_instance = True
+        fields = ('_id', 'name', 'price')
