+++ app/repositories/serializers/order_ingredient_serializer.py
@@ -1,0 +1,15 @@
+from app.plugins import ma
+from app.repositories.models.order_ingredient import OrderIngredient
+from .ingredient_serializer import IngredientSerializer
+
+class OrderIngredientSerializer(ma.SQLAlchemyAutoSchema):
+
+    ingredient = ma.Nested(IngredientSerializer)
+
+    class Meta:
+        model = OrderIngredient
+        load_instance = True
+        fields = (
+            'ingredient_price',
+            'ingredient'
+        )
